wrap entire app with MUI ThemeProvider
