+++ client/src/App.js
@@ -1,20 +1,45 @@
 import { Route, Routes } from 'react-router-dom';
+import { createTheme, ThemeProvider } from '@mui/material';
 import { Login, Signup, Landing, Dashboard, Account } from './pages';
 import axios from 'axios';
 
 axios.defaults.baseURL = process.env.REACT_APP_BACKEND_URL;
 
+const theme = createTheme({
+  palette: {
+    primary: {
+      main: '#896345',
+    },
+    secondary: {
+      main: '#f0eee7',
+    },
+  },
+  components: {
+    MuiFab: {
+      styleOverrides: {
+        root: {
+          position: 'fixed',
+          bottom: '20px',
+          right: '20px',
+        },
+      },
+    },
+  },
+});
+
 function App() {
   return (
-    <div className="App">
-      <Routes>
-        <Route path="/" element={<Landing />} />
-        <Route path="/dashboard" element={<Dashboard />} />
-        <Route path="/login" element={<Login />} />
-        <Route path="/signup" element={<Signup />} />
-        <Route path="/account" element={<Account />} />
-      </Routes>
-    </div>
+    <ThemeProvider theme={theme}>
+      <div className="App">
+        <Routes>
+          <Route path="/" element={<Landing />} />
+          <Route path="/dashboard" element={<Dashboard />} />
+          <Route path="/login" element={<Login />} />
+          <Route path="/signup" element={<Signup />} />
+          <Route path="/account" element={<Account />} />
+        </Routes>
+      </div>
+    </ThemeProvider>
   );
 }
