+++ client/src/pages/Login.jsx
@@ -4,19 +4,8 @@ import { Helmet } from 'react-helmet';
 import { ToastContainer } from 'react-toastify';
 import { Navbar, PasswordInputField } from 'components';
 import { isLoggedIn, loginUser, toastMessage } from 'helpers';
-import { createTheme, ThemeProvider, Button, Paper, TextField } from '@mui/material';
+import { Button, Paper, TextField } from '@mui/material';
 import { styled } from '@mui/material/styles';
-
-const theme = createTheme({
-  palette: {
-    primary: {
-      main: '#634832',
-    },
-    secondary: {
-      main: '#553E2B',
-    },
-  },
-});
 
 const LoginButton = styled((props) => <Button {...props} />)(({ theme }) => ({
   fontSize: 18,
@@ -79,47 +68,45 @@ function Login() {
         <title>Login | RoastRest</title>
       </Helmet>
       <Navbar />
-      <ThemeProvider theme={theme}>
-        <div className="max-w-screen-lg mx-auto px-3">
-          <Paper className="mx-auto mt-20 bg-white p-5 sm:w-2/3">
-            <h2 className="text-2xl font-semibold pb-2">Welcome to RoastRest!</h2>
-            <p className="text-md pb-2">Log back in to continue tracking your coffee beans.</p>
-            <form onSubmit={handleSubmit}>
-              <TextField
-                required
-                id="email"
-                name="email"
-                label="Email"
-                type="email"
-                value={email}
-                autoComplete="email"
-                onChange={handleOnChange}
-                margin="normal"
-                fullWidth
-              />
-              <PasswordInputField
-                id="password"
-                name="password"
-                label="Password"
-                value={password}
-                autoComplete="password"
-                handleOnChange={handleOnChange}
-                margin="normal"
-              />
-              <LoginButton type="submit" variant="contained" color="primary" fullWidth>
-                Login
-              </LoginButton>
-              <p className="mt-2">
-                Don't an account?{' '}
-                <Link className="font-semibold no-underline text-rr-brown-primary" to={'/signup'}>
-                  Sign up
-                </Link>
-              </p>
-            </form>
-          </Paper>
-          <ToastContainer />
-        </div>
-      </ThemeProvider>
+      <div className="max-w-screen-lg mx-auto px-3">
+        <Paper className="mx-auto mt-20 bg-white p-5 sm:w-2/3">
+          <h2 className="text-2xl font-semibold pb-2">Welcome to RoastRest!</h2>
+          <p className="text-md pb-2">Log back in to continue tracking your coffee beans.</p>
+          <form onSubmit={handleSubmit}>
+            <TextField
+              required
+              id="email"
+              name="email"
+              label="Email"
+              type="email"
+              value={email}
+              autoComplete="email"
+              onChange={handleOnChange}
+              margin="normal"
+              fullWidth
+            />
+            <PasswordInputField
+              id="password"
+              name="password"
+              label="Password"
+              value={password}
+              autoComplete="password"
+              handleOnChange={handleOnChange}
+              margin="normal"
+            />
+            <LoginButton type="submit" variant="contained" color="primary" fullWidth>
+              Login
+            </LoginButton>
+            <p className="mt-2">
+              Don't an account?{' '}
+              <Link className="font-semibold no-underline text-rr-brown-primary" to={'/signup'}>
+                Sign up
+              </Link>
+            </p>
+          </form>
+        </Paper>
+        <ToastContainer />
+      </div>
     </>
   );
 }
