+++ client/src/pages/Signup.jsx
@@ -4,19 +4,8 @@ import { Helmet } from 'react-helmet';
 import { ToastContainer } from 'react-toastify';
 import { Navbar, PasswordInputField } from 'components';
 import { registerUser, toastMessage } from 'helpers';
-import { createTheme, ThemeProvider, Button, Paper, TextField } from '@mui/material';
+import { Button, Paper, TextField } from '@mui/material';
 import { styled } from '@mui/material/styles';
-
-const theme = createTheme({
-  palette: {
-    primary: {
-      main: '#634832',
-    },
-    secondary: {
-      main: '#553E2B',
-    },
-  },
-});
 
 const SignupButton = styled((props) => <Button {...props} />)(({ theme }) => ({
   fontSize: 18,
@@ -72,67 +61,65 @@ function Signup() {
         <title>Sign Up | RoastRest</title>
       </Helmet>
       <Navbar />
-      <ThemeProvider theme={theme}>
-        <div className="max-w-screen-lg mx-auto px-3">
-          <Paper className="mx-auto mt-20 bg-white p-5 sm:w-2/3">
-            <h2 className="text-2xl font-semibold pb-2">Welcome to RoastRest!</h2>
-            <p className="text-md pb-2">Create an account and keep track of your coffee.</p>
-            <form autoComplete="off" onSubmit={handleSubmit}>
-              <TextField
-                required
-                id="name"
-                name="name"
-                label="Name"
-                value={name}
-                autoComplete="given-name"
-                onChange={handleOnChange}
-                margin="normal"
-                fullWidth
-              />
-              <TextField
-                required
-                id="email"
-                name="email"
-                label="Email"
-                type="email"
-                value={email}
-                autoComplete="email"
-                onChange={handleOnChange}
-                margin="normal"
-                fullWidth
-              />
-              <PasswordInputField
-                id="password"
-                name="password"
-                label="Password"
-                value={password}
-                autoComplete="new-password"
-                handleOnChange={handleOnChange}
-                margin="normal"
-              />
-              <PasswordInputField
-                id="confirmPassword"
-                name="confirmPassword"
-                label="Confirm password"
-                value={confirmPassword}
-                autoComplete="new-password"
-                handleOnChange={handleOnChange}
-                margin="normal"
-              />
-              <SignupButton type="submit" variant="contained" color="primary" fullWidth>
-                Sign Up
-              </SignupButton>
-              <p className="mt-2">
-                Already have an account?{' '}
-                <Link className="font-semibold no-underline text-rr-brown-primary" to={'/login'}>
-                  Sign in
-                </Link>
-              </p>
-            </form>
-          </Paper>
-          <ToastContainer />
-        </div>
-      </ThemeProvider>
+      <div className="max-w-screen-lg mx-auto px-3">
+        <Paper className="mx-auto mt-20 bg-white p-5 sm:w-2/3">
+          <h2 className="text-2xl font-semibold pb-2">Welcome to RoastRest!</h2>
+          <p className="text-md pb-2">Create an account and keep track of your coffee.</p>
+          <form autoComplete="off" onSubmit={handleSubmit}>
+            <TextField
+              required
+              id="name"
+              name="name"
+              label="Name"
+              value={name}
+              autoComplete="given-name"
+              onChange={handleOnChange}
+              margin="normal"
+              fullWidth
+            />
+            <TextField
+              required
+              id="email"
+              name="email"
+              label="Email"
+              type="email"
+              value={email}
+              autoComplete="email"
+              onChange={handleOnChange}
+              margin="normal"
+              fullWidth
+            />
+            <PasswordInputField
+              id="password"
+              name="password"
+              label="Password"
+              value={password}
+              autoComplete="new-password"
+              handleOnChange={handleOnChange}
+              margin="normal"
+            />
+            <PasswordInputField
+              id="confirmPassword"
+              name="confirmPassword"
+              label="Confirm password"
+              value={confirmPassword}
+              autoComplete="new-password"
+              handleOnChange={handleOnChange}
+              margin="normal"
+            />
+            <SignupButton type="submit" variant="contained" color="primary" fullWidth>
+              Sign Up
+            </SignupButton>
+            <p className="mt-2">
+              Already have an account?{' '}
+              <Link className="font-semibold no-underline text-rr-brown-primary" to={'/login'}>
+                Sign in
+              </Link>
+            </p>
+          </form>
+        </Paper>
+        <ToastContainer />
+      </div>
     </>
   );
 }
